perf: use pkg-types for package types
`read-package-up` pulls in ~400KB of unused types and functionality, thanks
to depending on `type-fest` - a bag of loosely related utilities.

We can drop this and use the much more focused `pkg-types` from unjs.

This does mean we no longer normalise `package.json`, but that was already
questionable since the normalisation is very specific to npm's internals (it
was built for npm, not for wide usage).

We can also lean on the already-installed empathic to find the `package.json`.